arreglados fallos de formato (int -> Float)

+++ app/src/main/java/com/azure/foodfinder/demasClases/Sub.kt
@@ -1,11 +1,11 @@
 package com.azure.foodfinder.demasClases
 
 data class Sub(
-    val daily: Int,
+    val daily: Float,
     val hasRDI: Boolean,
     val label: String,
     val schemaOrgTag: String,
     val tag: String,
-    val total: Int,
+    val total: Float,
     val unit: String
 )
